refactor: split queries.sql.ts into grouped files based on query functionality #355
